chore(ci): Build a macos wheel
SFCGAL is now able to generate an homebrew formulae and publish it to
the SFCGAL website.

This new ci job is able to use homebrew to download this formulae and
install the latest SFCGAL latest. This allows to build macos wheels
and then "repair" (add the dependencies) the wheel with
`delocate`. This means that the generated wheels can directly be used
without having CGAL or SFCGAL installed.

This is similar to what is already done for the linux and windows
wheels.

See: https://gitlab.com/sfcgal/SFCGAL/-/merge_requests/500