chore(ci): Generate wheel with dev version number
26ef8634bd076370cd27b66b1fd31def2d90c3a4 added support to upload the
wheels to gitlab package registry for the main version and on tags.

However, it is not possible to upload multiple times a new package
with the same version number. This issue is fixed by adding a date
based suffix to the version number. This allows to uniquify the
version number and to have increasing version numbers.

This achieved by adding the `increment_version.py` script. This script
is called when the commit does not have a tag. This changes the
version in the setup.py file and then builds the wheel as usual.

This works for the linux and windows wheels.