fix(sfcgal): Fix memory leak in is_valid_detail
`invalidity_reason` has been allocated on the C side by a call to
`strdup`, therefore it cannot be directly freed on the python
side. This is fixed by adding a call to `lib.sfcgal_free_buffer`

`invalidity_location` is not used and not allocated on the C side at
the moment. However, by adding a call `lib.sfcgal_geometry_delete`,
this prevents any memory leak if this is implemented in the future.