fix(sfcgal): Properly handle memory management on subgeometries
When a sub geometry is extracted from a geometry (for example a point
from a linestring), the parent geometry cannot be garbage collected as
long as the sub geometry is used because its memory is managed by the
parent geometry on the C++ side.

However, this is not properly managed nat the moment. For example, the
current code fails with a segmentation fault:

```python
point = LineString(((0, 0), (1, 0), (5, 2)))[-1]
wkt = point.to_wkt(1)
```

Indeed, on the first line a temporary linestring is created and a
point is immediately extracted. This ends up with a segmentation fault
because the linestring is immediately garbage collected but it owns
the generated point. Therefore, the point is also deleted on the C++
side when the linestring is garbage collected by the python
interpreter.

This issue is fixed by adding a `_parent` variable to a Geometry when
a sub geometry is created. This allows to keep a reference of the
parent geometry to prevent it from being garbage collected.