fix(update_def): Fix doxygen parser if block is not a function
The `try_to_parse_doxygen_block` function inspects the current line
and tries to parse the following lines if it's the start of a doxygen
block. It is needed to detect deprecated functions.

This parser needs to detect the end of the definition. It assumes that
a block ends with the definition of a C function and thus with `);`.

However, the C API header file starts with a lot of `typedef` which
doxygen blocks which end with `;`, not `);`. Therefore, those lines
are not properly parsed and the parser ends at the end of the first
function (`sfcgal_set_geometry_validation`) instead of the first
`typedef`.

It worked until recently because `sfcgal_set_geometry_validation` was
not deprecated. With this function being now deprecated, this function
is skipped from the generated c file. However, all the previous
`typedef` are also removed because they are wrongly part of the same
"doxygen block". This breaks pysfcgal build because some symbols are
missing.

This issue is fixed by checking the end of doxygen block with `;`, not
`);`. With this change, the `typedef` blocks are now properly parsed.